feat(project): add full feature complete server configuration menu
Added a fully working server configuration menu, including all the normal buttons available in
Portainer, some file management editing, as well as occasional status reports on the container.
Remove old system to download latest version of server jar file from each vendor's website, as the
docker image does that automatically so that I don't have to do that. Also removed some testing
files related to that since they are no longer needed. Add new test files relating to
containerization and starting the container (to make sure that it works), and changed a bit of
styling so that the status bubbles are distinct colors.